Bot, Player, Market, Bank added

+++ Capitalist.EXMPL/OBJECTS/ICapitalist.cs
@@ -6,7 +6,7 @@ public interface ICapitalist {
     public double Balance { get; set; }
     public Dictionary<string, int> Inventory { get; set; }
     public void Buy(string product, Dictionary<string, float> cost);
-    public void Sell(string product, Dictionary<string, float> cost);
+    public void Sell(string product, Dictionary<string, double> cost);
     public List<LoanOffer> LoanOffers { get; set; }
     public List<long> MyLoans { get; set; }
 }
